updated view to use float in the wait function

+++ auditory_stimulation/view/cli_view.py
@@ -21,5 +21,5 @@ class CLIView(AView):
         input("Please press the enter key")
         return True
 
-    def wait(self, secs: int) -> None:
+    def wait(self, secs: float) -> None:
         time.sleep(secs)
